IImage3d_GuiClient initial commit
This is a .Net-based GUI executable.  It allows the user to specify an image file and an IImage3d COM DLL that knows how to read that type of image.  It loads the image via the IImage3d DLL, and displays it in a 3-up plane view.

The intent is to provide a "neutral" testing environment to allow developers and testers to see the same behavior across sites and organizations - for example, "when I load image x with your IImage3d DLL, and navigate to slice y, I see a band of grey across the bottom.  Do you see what I mean?"  This behavior could then be replicated by a user at another site/company.