Auto-rearrange 2nd & 3rd axis returned by GetBoundingBox method
The DrawImages method assumes that the 1st axis is predominantly "X", 2nd is "Y" and 3rd "Z" in the bounding box. This is not necessarily always the case. The CVUS loader e.g. returns "Z" axis as 2nd and "Y" as 3rd. This change makes the TestViewer deal gracefully with this.