Add exception catching for version checking code
This code will throw if only a loader with non-matching bitness is installed, leading to a crash. This change will at least provide some information back to the user regarding the failure. Also, it's no longer fatal.

The best solution here would be to also search thought the part of the registry with non-matching bitness as SandboxTest already does. However, that can be improved later.