Update GetECG documentation
The doc string for GetECG previously stated that it should return S_OK with an empty ECG series if ECG is not available. This is problematic because it is unclear what an 'empty ECG Series' mean. What would be the start_time and delta_time of an empty ECG series be? Should the samples and trig_times be zero sized SAFEARRAYs, or should the by set to nullptr? This ambiguity can easily complicate client implementations and result in reliability issues if the client does not handle all situations correctly.

By returning E_FAIL from GetECG when there are no ECG, we avoid this problem. The down-side of this change is that we get failure codes in situations that may be considered normal execution. In .NET clients, error codes are translated into exceptions.