Update GetECG documentation to be explicit on the meaning of empty EcgSeries
The definition of an empty EcgSeries is that all struct members are default constructed, meaning that start_time and delta_time are set to 0.0, and that samples and trig_times are set to null pointers.

This avoids confusion around the meaning of empty EcgSeries, since this could be interpreted as having a zero size samples array. Now it is clear that an empty EcgSeries has a null pointer samples array.

The extension of this definition is that if the samples array is not null, it should at least contain one element. The same applies to the trig_times array.